refactor: move  projects components to Projects folder
