+++ React-Redux/Webpage/src/components/Projects/Clock/components/Display.js
