+++ React-Redux/Webpage/src/components/Projects/ContactManager/components/PeopleList.js
