+++ React-Redux/Webpage/src/components/Projects/Randomizer/App.js
