+++ React-Redux/Webpage/src/components/Projects/Randomizer/Randomizer.js
