+++ React-Redux/Webpage/src/components/Projects/Randomizer/features/modes/ModeButton.js
