+++ React-Redux/Webpage/src/components/Projects/Randomizer/features/results/Results.js
