+++ React-Redux/Webpage/src/components/Projects/Randomizer/store.js
