+++ React-Redux/Webpage/src/index.js
@@ -4,15 +4,15 @@ import ReactDOM from "react-dom";
 import "./style.scss";
 import reportWebVitals from "./reportWebVitals";
 
-import Navigation from "./components/Navigation";
-import Splash from "./components/Splash";
-import About from "./components/About";
-import Statement from "./components/Statement";
-import Skills from "./components/Skills";
-import Projects from "./components/Projects";
-import Gallery from "./components/Gallery";
-import Contact from "./components/Contact";
-import Footer from "./components/Footer";
+import Navigation from "./components/Navigation/Navigation";
+import Splash from "./components/Splash/Splash";
+import About from "./components/About/About";
+import Statement from "./components/Statement/Statement";
+import Skills from "./components/Skills/Skills";
+import Projects from "./components/Projects/Projects";
+import Gallery from "./components/Gallery/Gallery";
+import Contact from "./components/Contact/Contact";
+import Footer from "./components/Footer/Footer";
 
 ReactDOM.render(
   <React.StrictMode>
